Fix remaining import cycles
