+++ packages/react-native-sortable/src/components/index.ts
@@ -1,3 +1,3 @@
-export { SortableLayer } from './shared';
+export { createSortableTouchable, SortableLayer } from './shared';
 export { default as SortableFlex } from './SortableFlex';
 export { default as SortableGrid } from './SortableGrid';
