+++ packages/react-native-sortable/src/components/shared/index.ts
@@ -1,3 +1,4 @@
+export { default as createSortableTouchable } from './createSortableTouchable';
 export { default as DraggableView } from './DraggableView';
 export { default as SortableContainer } from './SortableContainer';
 export { default as SortableLayer } from './SortableLayer';
