--- packages/react-native-sortable/src/utils/hoc/index.ts
@@ -1,1 +1,0 @@
-export { default as createSortableTouchable } from './createSortableTouchable';
